+++ src/main/java/com/contentful/java/cda/interceptor/AuthorizationHeaderInterceptor.java
@@ -1,17 +1,10 @@
 package com.contentful.java.cda.interceptor;
 
-import java.io.IOException;
-
-import okhttp3.Interceptor;
-import okhttp3.Request;
-import okhttp3.Response;
-
 /**
  * Interceptor to add authorization header to requests
  */
-public class AuthorizationHeaderInterceptor implements Interceptor {
+public class AuthorizationHeaderInterceptor extends HeaderInterceptor {
   public static final String HEADER_NAME = "Authorization";
-  private final String token;
 
   /**
    * Create Header interceptor, saving parameters.
@@ -19,21 +12,7 @@ public class AuthorizationHeaderInterceptor implements Interceptor {
    * @param token the access token to be used with *every* request.
    */
   public AuthorizationHeaderInterceptor(String token) {
-    this.token = token;
+    super(HEADER_NAME, "Bearer " + token);
   }
 
-  /**
-   * Method called by framework, to enrich current request chain with the header information requested.
-   *
-   * @param chain the execution chain for the request.
-   * @return the response received.
-   * @throws IOException in case of failure down the line.
-   */
-  @Override public Response intercept(Chain chain) throws IOException {
-    final Request request = chain.request();
-
-    return chain.proceed(request.newBuilder()
-        .addHeader(HEADER_NAME, "Bearer " + token)
-        .build());
-  }
 }