+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Version [7.6.0][unreleased] - (in development)
 - Added: [Image API](https://www.contentful.com/developers/docs/references/images-api/#/introduction)
+- Added: Custom Contentful HTTP Header.
 
 ## Version [7.5.0] - (2017-03-30)
 - Added: Retrieve default call.factory from client builder.